requirements.txt
asgiref==3.5.2
Babel==2.10.3
certifi==2022.6.15
charset-normalizer==2.1.0
cloudinary==1.29.0
dj-database-url==0.5.0
Django==4.0.6
django-cloudinary-storage==0.3.0
django-heroku==0.3.1
django-phonenumber-field==6.3.0
gunicorn==20.1.0
idna==3.3
phonenumberslite==8.12.51
Pillow==9.2.0
psycopg2==2.9.3
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
whitenoise==6.2.0
